ai_filter/requirements.txt
numpy==1.26.4
kafka-python==2.0.2
prometheus-client==0.18.0
torch==2.2.2
transformers==4.41.2
sentence-transformers==2.7.0
detoxify